peak_detection to peak_det refactoring fix
